one_to_many

+++ app/Http/Controllers/TasksController.php
@@ -15,13 +15,21 @@ class TasksController extends Controller
      */
     public function index()
     {
-        // タスク一覧を取得
-        $tasks = Task::all();
-        
-        // タスク一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザのタスクの一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
     }
 
     /**
@@ -55,11 +63,17 @@ class TasksController extends Controller
         
         // タスクを作成
         $task = new Task;
-        dd(\Auth::id());
-        //$task->user_id = auth::;
+        //dd(\Auth::id());
+        $task->user_id = \Auth::id();
         $task->content = $request->content;
         $task->status = $request->status;
         $task->save();
+
+        // 認証済みユーザ（閲覧者）のタスクとして作成（リクエストされた値をもとに作成）
+        //$request->user()->tasks()->create([
+        //    'content' => $request->content,
+        //    'status' => $request->status,
+        //]);
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -135,10 +149,17 @@ class TasksController extends Controller
     {
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
+
         // タスクを削除
-        $task->delete();
+        //$task->delete();
+
+        // 認証済みユーザ（閲覧者）がそのタスクの所有者である場合は、タスクを削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
+        
     }
 }
